Move adapters into ./adapters subdirectory in ws-manager
